.h -> .hpp

+++ common/logging.cpp
@@ -1,7 +1,7 @@
-#include "logging.h"
+#include "logging.hpp"
 
-#include "console.h"
-#include "errutils.h"
+#include "console.hpp"
+#include "errutils.hpp"
 
 #ifdef HAVE_BOOST
 #include <boost/filesystem.hpp>
@@ -24,6 +24,8 @@
 namespace aifil {
 
 extern CONSOLE_STATE console_state;
+static FILE *printf_stream = stderr;
+static int required_log_level = LOG_NO_SPAM;
 
 #ifdef HAVE_BOOST
 class SimpleLogger
@@ -85,15 +87,15 @@ void SimpleLogger::message(const char* msg)
 	{
 		std::string fn = (dir_log + "/" + log_name);
 		file_all = fopen(fn.c_str(), "wb");
-		if (!file_all)
-			fprintf(stderr, "cannot open '%s'\n", fn.c_str());
+		if (!file_all && printf_stream)
+			fprintf(printf_stream, "\ncannot open '%s'\n", fn.c_str());
 		else
 		{
 #ifdef WIN32
 			int fdesc = _fileno(file_all);
 			HANDLE fhan = (HANDLE) _get_osfhandle(fdesc);
-			if (!SetHandleInformation(fhan, HANDLE_FLAG_INHERIT, 0))
-				fprintf(stderr, "Failed to set handle information\n");
+			if (!SetHandleInformation(fhan, HANDLE_FLAG_INHERIT, 0) && printf_stream)
+				fprintf(printf_stream, "Failed to set handle information\n");
 #endif
 			setbuf(file_all, 0);
 			fwrite(version.c_str(), version.size(), 1, file_all);
@@ -127,7 +129,8 @@ void SimpleLogger::message(const char* msg)
 //				true);
 //		} catch (const std::exception& e) {
 //			rotation_broken = true;
-//			fprintf(stderr, "Failed to rotate log: %s\n", e.what());
+//			if (printf_stream)
+//				fprintf(printf_stream, "Failed to rotate log: %s\n", e.what());
 //		}
 	}
 }
@@ -141,8 +144,10 @@ struct SimpleLogger
 static SimpleLogger *logger = 0;
 #endif
 
-void log_state(const char* fmt, ...)
+void log_debug(const char* fmt, ...)
 {
+	if (!(required_log_level & LOG_DEBUG))
+		return;
 	static char prefix[] = "\n";
 	char buf[65536];
 	strcpy(buf, prefix);
@@ -151,20 +156,56 @@ void log_state(const char* fmt, ...)
 	vsnprintf(buf + sizeof(prefix) - 1, sizeof(buf) - sizeof(prefix), fmt, ap);
 	va_end(ap);
 	buf[65535] = 0;
-	pretty_printf(LOG_COLOR_GRAY, stderr, buf);
+	pretty_printf(LOG_COLOR_GRAY, printf_stream, buf);
 	if (logger)
 		logger->message(buf);
 }
 
 void log_raw(const char *fmt, ...)
 {
+	if (!(required_log_level & LOG_RAW))
+		return;
 	char buf[65536];
 	va_list ap;
 	va_start(ap, fmt);
 	vsnprintf(buf, sizeof(buf), fmt, ap);
 	va_end(ap);
 	buf[65535] = 0;
-	pretty_printf(LOG_COLOR_GRAY, stderr, buf);
+	pretty_printf(LOG_COLOR_GRAY, printf_stream, buf);
+	if (logger)
+		logger->message(buf);
+}
+
+void log_state(const char* fmt, ...)
+{
+	if (!(required_log_level & LOG_STATE))
+		return;
+	static char prefix[] = "\n";
+	char buf[65536];
+	strcpy(buf, prefix);
+	va_list ap;
+	va_start(ap, fmt);
+	vsnprintf(buf + sizeof(prefix) - 1, sizeof(buf) - sizeof(prefix), fmt, ap);
+	va_end(ap);
+	buf[65535] = 0;
+	pretty_printf(LOG_COLOR_NORMAL, printf_stream, buf);
+	if (logger)
+		logger->message(buf);
+}
+
+void log_important(const char* fmt, ...)
+{
+	if (!(required_log_level & LOG_IMPORTANT))
+		return;
+	static char prefix[] = "\n";
+	char buf[65536];
+	strcpy(buf, prefix);
+	va_list ap;
+	va_start(ap, fmt);
+	vsnprintf(buf + sizeof(prefix) - 1, sizeof(buf) - sizeof(prefix), fmt, ap);
+	va_end(ap);
+	buf[65535] = 0;
+	pretty_printf(LOG_COLOR_HL, printf_stream, buf);
 	if (logger)
 		logger->message(buf);
 }
@@ -187,6 +228,9 @@ void log_quiet(const char* fmt, ...)
 
 void log_warning(const char* fmt, ...)
 {
+	if (!(required_log_level & LOG_WARNING))
+		return;
+
 	static char prefix[] = "\nWARNING: ";
 	char buf[65536];
 	strcpy(buf, prefix);
@@ -196,13 +240,16 @@ void log_warning(const char* fmt, ...)
 	vsnprintf(buf + sizeof(prefix) - 1, sizeof(buf) - sizeof(prefix), fmt, ap);
 	va_end(ap);
 	buf[65535] = 0;
-	pretty_printf(LOG_COLOR_YELLOW, stderr, buf);
+	pretty_printf(LOG_COLOR_YELLOW, printf_stream, buf);
 	if (logger)
 		logger->message(buf);
 }
 
 void log_error(const char* fmt, ...)
 {
+	if (!(required_log_level & LOG_ERROR))
+		return;
+
 	static char prefix[] = "\nERROR: ";
 	char buf[65536];
 	strcpy(buf, prefix);
@@ -212,13 +259,18 @@ void log_error(const char* fmt, ...)
 	vsnprintf(buf + sizeof(prefix) - 1, sizeof(buf) - sizeof(prefix), fmt, ap);
 	va_end(ap);
 	buf[65535] = 0;
-	pretty_printf(LOG_COLOR_RED, stderr, buf);
+	pretty_printf(LOG_COLOR_RED, printf_stream, buf);
+	if (printf_stream)
+		fflush(printf_stream);
 	if (logger)
 		logger->message(buf);
 }
 
 void log_ok(const char* fmt, ...)
 {
+	if (!(required_log_level & LOG_OK))
+		return;
+
 	static char prefix[] = "\n";
 	char buf[65536];
 	strcpy(buf, prefix);
@@ -228,23 +280,30 @@ void log_ok(const char* fmt, ...)
 	vsnprintf(buf + sizeof(prefix) - 1, sizeof(buf) - sizeof(prefix), fmt, ap);
 	va_end(ap);
 	buf[65535] = 0;
-	pretty_printf(LOG_COLOR_GREEN, stderr, buf);
+	pretty_printf(LOG_COLOR_GREEN, printf_stream, buf);
 	if (logger)
 		logger->message(buf);
 }
 
-
-void log_state(const std::string & msg)
+void log_debug(const std::string & msg)
 {
-	log_state(msg.c_str());
+	log_debug(msg.c_str());
 }
-
 
 void log_raw(const std::string & msg)
 {
 	log_raw(msg.c_str());
 }
 
+void log_state(const std::string & msg)
+{
+	log_state(msg.c_str());
+}
+
+void log_important(const std::string &msg)
+{
+	log_important(msg.c_str());
+}
 
 void log_warning(const std::string & msg)
 {
@@ -265,40 +324,72 @@ void log_ok(const std::string & msg)
 
 
 #ifdef HAVE_BOOST
-void logger_startup(const std::string& directory_to_write_logs,
-	const std::string& directory_to_save_crashlogs,
-	const std::string& logfile_name)
+
+// TODO Сделать разбор путей с учетом наличия/отсутствия слэшей.
+
+void logger_startup(
+		FILE *stream_for_printf,
+		int log_level,
+		const std::string & directory_to_write_logs,
+		const std::string & directory_to_save_crashlogs,
+		const std::string & logfile_name)
 {
+	printf_stream = stream_for_printf;
+	required_log_level = log_level;
+	if (!printf_stream)
+	{
+		printf("SILENT MODE! It is the last message, now your app "
+			"will not redirect any console output anywhere!\n");
+	}
+
+	if (directory_to_write_logs.empty())
+		return;
+
 	if (logger)
 	{
-		printf("Logger is already set\n");
+		printf("\nLogger is already set\n");
 		return;
 	}
+
 	af_assert(!directory_to_write_logs.empty() && !directory_to_save_crashlogs.empty());
-	try {
+
+	try
+	{
 		boost::filesystem::path t(logfile_name);
 		const std::string log_stem = t.stem().string();
 		const std::string log_extension = t.extension().string();
 
-		const std::string log = directory_to_write_logs + "/" + logfile_name;
-		const std::string log_bkp = directory_to_save_crashlogs + "/" + log_stem
-			+ ".bkp" + log_extension;
+//		const std::string log = directory_to_write_logs + "/" + logfile_name;
+//		const std::string log_bkp = directory_to_save_crashlogs + "/" + log_stem
+//			+ ".bkp" + log_extension;
 
 		boost::filesystem::remove(directory_to_write_logs + "/" + log_stem + ".older" + log_extension);
-		logger.reset(new SimpleLogger(directory_to_write_logs, log));
+		logger.reset(new SimpleLogger(directory_to_write_logs, logfile_name));
 
-	} catch (const std::exception& e) {
-		printf("LOGGER ERROR: %s\n", e.what());
+	}
+
+	catch (const std::exception& e)
+	{
+		printf("\nLOGGER ERROR: %s\n", e.what());
 	}
 }
+
 
 void logger_shutdown()
 {
 	// Removed
 	// std::unique_lock<std::mutex> lock(logging_mutex);
 
+	if (printf_stream)
+		fprintf(printf_stream, "\n");
 	logger.reset();
 }
+
+void logger_set_mode(int log_level)
+{
+	required_log_level = log_level;
+}
+
 #else
 
 void logger_startup(const std::string& directory_to_write_logs,
